Create separate files for storing attribute definitions & aggregate definitions for occurrences on EAP
